+++ changelog.md
@@ -4,6 +4,7 @@ Upcoming Release (TBD)
 Bug Fixes:
 ----------
 
+* fix SSL through SSH jump host by using a true python socket for a tunnel
 
 Internal:
 ---------